uint32 instead of uint16 in simple net

+++ dqn/gym/net.py
@@ -111,7 +111,7 @@ class SimpleNet:
                     tf.float32, shape=[None, convnet_pars['n_approximators']])
 
             if convnet_pars['n_states'] is not None:
-                x = tf.one_hot(tf.cast(self._x[..., 0], tf.uint16),
+                x = tf.one_hot(tf.cast(self._x[..., 0], tf.uint32),
                                convnet_pars['n_states'], axis=1)[..., 0]
             else:
                 x = self._x[..., 0]
